Add files via upload
Predicting Customer Clicked Ads Classification Using Machine Learning

A company in Indonesia aims to determine the effectiveness of an advertisement they have launched. This is crucial for a company operating in the digital marketing consulting field to gauge the reach of their marketed ads and attract customers to view the ads. The company seeks to extract insights regarding user behavior from the collected data by creating visualizations, developing a relevant machine learning model to meet their needs, and providing recommendations based on the findings.